Refs #9: Setting and tasks for shared virtualenvs
Adds a setting, :shared_virtualenv, which causes the deployment process
to check for and create a virtualenv in the shared_path, and then
symlink it into the release_path.  This reduces deploy time in the
optimal case (requirements haven't changed) by however long it takes to
build requirements, which can be substantial.